fix #264594 lyrics update
